port History and Map Activities to Fragments
