+++ app/src/androidTest/java/ch/epfl/sdp/HistoryActivityTest.java
@@ -49,6 +49,8 @@ public class HistoryActivityTest {
     @Rule
     public MockitoRule mockitoRule = MockitoJUnit.rule();
 
+    private HistoryFragment fragment;
+
     @Before
     public void setupMockito() {
         when(querySnapshot.iterator()).thenReturn(Collections.singletonList(queryDocumentSnapshot).iterator());
@@ -59,6 +61,8 @@ public class HistoryActivityTest {
         when(unreadableSnapshot.iterator()).thenReturn(Collections.singletonList(unreadableDocumentSnapshot).iterator());
         when(unreadableDocumentSnapshot.get("Time")).thenReturn(null);
         when(unreadableDocumentSnapshot.get("Position")).thenReturn(new GeoPoint(19, 98));
+
+        fragment = (HistoryFragment) mActivityRule.getActivity().getSupportFragmentManager().findFragmentById(R.id.history_fragment);
     }
 
     @Test
@@ -70,7 +74,7 @@ public class HistoryActivityTest {
             }
         };
 
-        mActivityRule.getActivity().setFirestoreInteractor(successInteractor);
+        fragment.setFirestoreInteractor(successInteractor);
 
         onView(withId(R.id.refresh_history)).perform(click());
         onView(withId(R.id.conn_status)).check(matches(withText("QUERY OK")));
@@ -97,7 +101,7 @@ public class HistoryActivityTest {
             }
         };
 
-        mActivityRule.getActivity().setFirestoreInteractor(failureInteractor);
+        fragment.setFirestoreInteractor(failureInteractor);
 
         onView(withId(R.id.refresh_history)).perform(click());
         onView(withId(R.id.conn_status)).check(matches(withText("CONNECTION ERROR")));
@@ -112,7 +116,7 @@ public class HistoryActivityTest {
             }
         };
 
-        mActivityRule.getActivity().setFirestoreInteractor(unreadableInteractor);
+        fragment.setFirestoreInteractor(unreadableInteractor);
 
         onView(withId(R.id.refresh_history)).perform(click());
         onData(anything())
